Finish Session partial class files Start re-writing tests so they are not dependant on any existing objects that are not there by default.
