--- src/Session/AAAAAA.cs
@@ -1,14 +1,0 @@
-﻿using Koopman.CheckPoint.Common;
-using Koopman.CheckPoint.Internal;
-using Koopman.CheckPoint.Json;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Linq;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-
-namespace Koopman.CheckPoint
-{
-    public partial class Session
-    {
-    }
-}
