+++ src/Session/ServiceICMP6.cs
@@ -1,0 +1,194 @@
+﻿using Koopman.CheckPoint.Common;
+using Koopman.CheckPoint.Internal;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace Koopman.CheckPoint
+{
+    public partial class Session
+    {
+        #region Methods
+
+        /// <summary>
+        /// Deletes a service-icmp6.
+        /// </summary>
+        /// <param name="value">The name or UID to delete.</param>
+        /// <param name="ignore">Weather warnings or errors should be ignored</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns></returns>
+        public Task DeleteServiceICMP6
+            (
+                string value,
+                Ignore ignore = Delete.Defaults.ignore,
+                CancellationToken cancellationToken = default
+            )
+        {
+            return Delete.Invoke
+                (
+                    Session: this,
+                    Command: "delete-service-icmp6",
+                    Value: value,
+                    Ignore: ignore,
+                    cancellationToken: cancellationToken
+                );
+        }
+
+        /// <summary>
+        /// Finds all service-icmp6s that match filter.
+        /// </summary>
+        /// <param name="filter">The filter.</param>
+        /// <param name="ipOnly">
+        /// if set to <c>true</c> will search objects by their IP address only, without involving the
+        /// textual search.
+        /// </param>
+        /// <param name="detailLevel">The detail level.</param>
+        /// <param name="limit">The limit.</param>
+        /// <param name="order">The order.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>Array of Objects</returns>
+        public Task<ServiceICMP6[]> FindAllServicesICMP6
+            (
+                string filter,
+                bool ipOnly = FindAll.Defaults.IPOnly,
+                DetailLevels detailLevel = FindAll.Defaults.DetailLevel,
+                int limit = FindAll.Defaults.Limit,
+                IOrder order = FindAll.Defaults.Order,
+                CancellationToken cancellationToken = default
+            )
+        {
+            return FindAll.Invoke<ServiceICMP6>
+                (
+                    Session: this,
+                    Type: "service-icmp6",
+                    Filter: filter,
+                    IPOnly: ipOnly,
+                    DetailLevel: detailLevel,
+                    Limit: limit,
+                    Order: order,
+                    cancellationToken: cancellationToken
+                );
+        }
+
+        /// <summary>
+        /// Finds all services ICMP6.
+        /// </summary>
+        /// <param name="detailLevel">The detail level to return.</param>
+        /// <param name="limit">The limit.</param>
+        /// <param name="order">The order.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>Array of Objects</returns>
+        public Task<ServiceICMP6[]> FindAllServicesICMP6
+            (
+                DetailLevels detailLevel = FindAll.Defaults.DetailLevel,
+                int limit = FindAll.Defaults.Limit,
+                IOrder order = FindAll.Defaults.Order,
+                CancellationToken cancellationToken = default
+            )
+        {
+            return FindAll.Invoke<ServiceICMP6>
+                (
+                    Session: this,
+                    Command: "show-services-icmp6",
+                    DetailLevel: detailLevel,
+                    Limit: limit,
+                    Order: order,
+                    cancellationToken: cancellationToken
+                );
+        }
+
+        /// <summary>
+        /// Finds a service-icmp6.
+        /// </summary>
+        /// <param name="value">The name or UID to find.</param>
+        /// <param name="detailLevel">The detail level of child objects to return.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>Object object</returns>
+        public Task<ServiceICMP6> FindServiceICMP6
+            (
+                string value,
+                DetailLevels detailLevel = Find.Defaults.DetailLevel,
+                CancellationToken cancellationToken = default
+            )
+        {
+            return Find.Invoke<ServiceICMP6>
+                (
+                    Session: this,
+                    Command: "show-service-icmp6",
+                    Value: value,
+                    DetailLevel: detailLevel,
+                    cancellationToken: cancellationToken
+                );
+        }
+
+        /// <summary>
+        /// Finds service-icmp6s that match filter.
+        /// </summary>
+        /// <param name="filter">The filter.</param>
+        /// <param name="ipOnly">
+        /// if set to <c>true</c> will search objects by their IP address only, without involving the
+        /// textual search.
+        /// </param>
+        /// <param name="detailLevel">The detail level.</param>
+        /// <param name="limit">The limit.</param>
+        /// <param name="offset">The offset.</param>
+        /// <param name="order">The order.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>NetworkObjectsPagingResults of Objects</returns>
+        public Task<NetworkObjectsPagingResults<ServiceICMP6>> FindServicesICMP6
+            (
+                string filter,
+                bool ipOnly = Finds.Defaults.IPOnly,
+                DetailLevels detailLevel = Finds.Defaults.DetailLevel,
+                int limit = Finds.Defaults.Limit,
+                int offset = Finds.Defaults.Offset,
+                IOrder order = Finds.Defaults.Order,
+                CancellationToken cancellationToken = default
+            )
+        {
+            return Finds.Invoke<ServiceICMP6>
+                (
+                    Session: this,
+                    Type: "service-icmp6",
+                    Filter: filter,
+                    IPOnly: ipOnly,
+                    DetailLevel: detailLevel,
+                    Limit: limit,
+                    Offset: offset,
+                    Order: order,
+                    cancellationToken: cancellationToken
+                );
+        }
+
+        /// <summary>
+        /// Finds services ICMP6.
+        /// </summary>
+        /// <param name="detailLevel">The detail level to return.</param>
+        /// <param name="limit">The limit.</param>
+        /// <param name="offset">The offset.</param>
+        /// <param name="order">The order.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>NetworkObjectsPagingResults of Objects</returns>
+        public Task<NetworkObjectsPagingResults<ServiceICMP6>> FindServicesICMP6
+            (
+                DetailLevels detailLevel = Finds.Defaults.DetailLevel,
+                int limit = Finds.Defaults.Limit,
+                int offset = Finds.Defaults.Offset,
+                IOrder order = Finds.Defaults.Order,
+                CancellationToken cancellationToken = default
+            )
+        {
+            return Finds.Invoke<ServiceICMP6>
+                (
+                    Session: this,
+                    Command: "show-services-icmp6",
+                    DetailLevel: detailLevel,
+                    Limit: limit,
+                    Offset: offset,
+                    Order: order,
+                    cancellationToken: cancellationToken
+                );
+        }
+
+        #endregion Methods
+    }
+}
