+++ src/Session/ServiceUDP.cs
@@ -1,0 +1,194 @@
+﻿using Koopman.CheckPoint.Common;
+using Koopman.CheckPoint.Internal;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace Koopman.CheckPoint
+{
+    public partial class Session
+    {
+        #region Methods
+
+        /// <summary>
+        /// Deletes a UDP service.
+        /// </summary>
+        /// <param name="value">The name or UID to delete.</param>
+        /// <param name="ignore">Weather warnings or errors should be ignored</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns></returns>
+        public Task DeleteServiceUDP
+            (
+                string value,
+                Ignore ignore = Delete.Defaults.ignore,
+                CancellationToken cancellationToken = default
+            )
+        {
+            return Delete.Invoke
+                (
+                    Session: this,
+                    Command: "delete-service-udp",
+                    Value: value,
+                    Ignore: ignore,
+                    cancellationToken: cancellationToken
+                );
+        }
+
+        /// <summary>
+        /// Finds all UDP services that match filter.
+        /// </summary>
+        /// <param name="filter">The filter.</param>
+        /// <param name="ipOnly">
+        /// if set to <c>true</c> will search objects by their IP address only, without involving the
+        /// textual search.
+        /// </param>
+        /// <param name="detailLevel">The detail level.</param>
+        /// <param name="limit">The limit.</param>
+        /// <param name="order">The order.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>Array of ServiceTCP</returns>
+        public Task<ServiceUDP[]> FindAllServicesUDP
+            (
+                string filter,
+                bool ipOnly = FindAll.Defaults.IPOnly,
+                DetailLevels detailLevel = FindAll.Defaults.DetailLevel,
+                int limit = FindAll.Defaults.Limit,
+                IOrder order = FindAll.Defaults.Order,
+                CancellationToken cancellationToken = default
+            )
+        {
+            return FindAll.Invoke<ServiceUDP>
+                (
+                    Session: this,
+                    Type: "service-udp",
+                    Filter: filter,
+                    IPOnly: ipOnly,
+                    DetailLevel: detailLevel,
+                    Limit: limit,
+                    Order: order,
+                    cancellationToken: cancellationToken
+                );
+        }
+
+        /// <summary>
+        /// Finds all UDP services.
+        /// </summary>
+        /// <param name="detailLevel">The detail level to return.</param>
+        /// <param name="limit">The limit.</param>
+        /// <param name="order">The order.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>Array of ServiceTCP</returns>
+        public Task<ServiceUDP[]> FindAllServicesUDP
+            (
+                DetailLevels detailLevel = FindAll.Defaults.DetailLevel,
+                int limit = FindAll.Defaults.Limit,
+                IOrder order = FindAll.Defaults.Order,
+                CancellationToken cancellationToken = default
+            )
+        {
+            return FindAll.Invoke<ServiceUDP>
+                (
+                    Session: this,
+                    Command: "show-services-udp",
+                    DetailLevel: detailLevel,
+                    Limit: limit,
+                    Order: order,
+                    cancellationToken: cancellationToken
+                );
+        }
+
+        /// <summary>
+        /// Finds UDP services that match filter.
+        /// </summary>
+        /// <param name="filter">The filter.</param>
+        /// <param name="ipOnly">
+        /// if set to <c>true</c> will search objects by their IP address only, without involving the
+        /// textual search.
+        /// </param>
+        /// <param name="detailLevel">The detail level.</param>
+        /// <param name="limit">The limit.</param>
+        /// <param name="offset">The offset.</param>
+        /// <param name="order">The order.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>NetworkObjectsPagingResults of ServiceTCP</returns>
+        public Task<NetworkObjectsPagingResults<ServiceUDP>> FindServicesUDP
+            (
+                string filter,
+                bool ipOnly = Finds.Defaults.IPOnly,
+                DetailLevels detailLevel = Finds.Defaults.DetailLevel,
+                int limit = Finds.Defaults.Limit,
+                int offset = Finds.Defaults.Offset,
+                IOrder order = Finds.Defaults.Order,
+                CancellationToken cancellationToken = default
+            )
+        {
+            return Finds.Invoke<ServiceUDP>
+                (
+                    Session: this,
+                    Type: "service-udp",
+                    Filter: filter,
+                    IPOnly: ipOnly,
+                    DetailLevel: detailLevel,
+                    Limit: limit,
+                    Offset: offset,
+                    Order: order,
+                    cancellationToken: cancellationToken
+                );
+        }
+
+        /// <summary>
+        /// Finds UDP services.
+        /// </summary>
+        /// <param name="detailLevel">The detail level to return.</param>
+        /// <param name="limit">The limit.</param>
+        /// <param name="offset">The offset.</param>
+        /// <param name="order">The order.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>NetworkObjectsPagingResults of ServiceTCP</returns>
+        public Task<NetworkObjectsPagingResults<ServiceUDP>> FindServicesUDP
+            (
+                DetailLevels detailLevel = Finds.Defaults.DetailLevel,
+                int limit = Finds.Defaults.Limit,
+                int offset = Finds.Defaults.Offset,
+                IOrder order = Finds.Defaults.Order,
+                CancellationToken cancellationToken = default
+            )
+        {
+            return Finds.Invoke<ServiceUDP>
+                (
+                    Session: this,
+                    Command: "show-services-udp",
+                    DetailLevel: detailLevel,
+                    Limit: limit,
+                    Offset: offset,
+                    Order: order,
+                    cancellationToken: cancellationToken
+                );
+        }
+
+        /// <summary>
+        /// Finds a UDP service.
+        /// </summary>
+        /// <param name="value">The name or UID to find.</param>
+        /// <param name="detailLevel">The detail level of child objects to return.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>ServiceUDP object</returns>
+        public Task<ServiceUDP> FindServiceUDP
+            (
+                string value,
+                DetailLevels detailLevel = Find.Defaults.DetailLevel,
+                CancellationToken cancellationToken = default
+            )
+        {
+            return Find.Invoke<ServiceUDP>
+                (
+                    Session: this,
+                    Command: "show-service-udp",
+                    Value: value,
+                    DetailLevel: detailLevel,
+                    cancellationToken: cancellationToken
+                );
+        }
+
+        #endregion Methods
+    }
+}
